requirements.txt
asgiref==3.7.2
attrs==23.1.0
coverage==7.3.1
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
drf-extensions==0.7.1
drf-spectacular==0.25.1
drf-spectacular-jsonapi==0.3.2
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mysqlclient==2.2.0
packaging==23.1
pluggy==1.3.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.2
sqlparse==0.4.4
uritemplate==4.1.1
